+++ vllm/engine/llm_engine.py
@@ -7,8 +7,8 @@ from vllm.config import (CacheConfig, DeviceConfig, ModelConfig,
                          ParallelConfig, SchedulerConfig, LoRAConfig)
 from vllm.core.scheduler import Scheduler, SchedulerOutputs
 from vllm.engine.arg_utils import EngineArgs
-from vllm.executor.single_gpu_executor import GPUExecutor
-from vllm.executor.ray_distributed_executor import RayGPUExecutor
+from vllm.executor.gpu_executor import GPUExecutor
+from vllm.executor.ray_gpu_executor import RayGPUExecutor
 from vllm.engine.metrics import StatLogger, Stats
 from vllm.engine.ray_utils import initialize_ray_cluster
 from vllm.logger import init_logger
